+++ angular/lib/src/facade/lang.dart
@@ -15,26 +15,6 @@ List<String> jsSplit(String s, RegExp regExp) {
   return parts;
 }
 
-// Dart VM implements `identical` as true reference identity. JavaScript does
-// not have this. The closest we have in JS is `===`. However, for strings JS
-// would actually compare the contents rather than references. `dart2js`
-// compiles `identical` to `===` and therefore there is a discrepancy between
-// Dart VM and `dart2js`. The implementation of `looseIdentical` attempts to
-// bridge the gap between the two while retaining good performance
-// characteristics. In JS we use simple `identical`, which compiles to `===`,
-// and in Dart VM we emulate the semantics of `===` by special-casing strings.
-// Note that the VM check is a compile-time constant. This allows `dart2js` to
-// evaluate the conditional during compilation and inline the entire function.
-//
-// See: dartbug.com/22496, dartbug.com/25270
-const isDartVM = !identical(1.0, 1); // a hack
-bool looseIdentical(a, b) => isDartVM ? _looseIdentical(a, b) : identical(a, b);
-
-// This function is intentionally separated from `looseIdentical` to keep the
-// number of AST nodes low enough for `dart2js` to inline the code.
-bool _looseIdentical(a, b) =>
-    a is String && b is String ? a == b : identical(a, b);
-
 /// Use this function to guard debugging code. When Dart is compiled in
 /// production mode, the code guarded using this function will be tree
 /// shaken away, reducing code size.